app/models/user.py
from tortoise import fields, models
from tortoise.contrib.pydantic import pydantic_model_creator


class User(models.Model):
    full_name = fields.TextField()
    email = fields.TextField()
    created_at = fields.DatetimeField(auto_now_add=True)

    def __str__(self):
        return self.email


UserSchema = pydantic_model_creator(User)
