pyproject.toml
[tool.poetry]
name = "fileupload-fastapi"
version = "0.1.0"
description = "A backend for simple file uploads"
authors = ["Andrew Yourkin <andrew@yourkin.io>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.2"
uvicorn = "^0.14.0"
asyncpg = "^0.23.0"
tortoise-orm = "^0.17.4"
aerich = "^0.5.3"
pytest = "^6.2.4"
requests = "^2.25.1"
python-multipart = "^0.0.5"
aiofiles = "^0.7.0"
httpx = "^0.18.2"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
mypy = "^0.902"
types-aiofiles = "^0.1.5"
pytest-mypy = "^0.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
